requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.9.2
pydantic-settings==2.3.0
httpx==0.27.0
python-dotenv==1.0.1
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt<4.0.0
python-multipart==0.0.9